Refactor dashboard navigation into DashboardNavigationCoordinator
Extracted initial fragment routing, notification intent routing, and tab selection logic from DashboardActivity and DashboardElementActivity into a new DashboardNavigationCoordinator. This improves code organization while maintaining existing repository calls to limit the refactoring scope.
